+++ thunderstore_package/CHANGELOG.md
@@ -1,5 +1,9 @@
+# v0.2.1
+ - Configuration keys for items rarity
+ - Rebalanced item rarity
+
 # v0.2.0
- - Fix carton format quantity
+ - Fixed carton format quantity
  - Added tomato juice (Saydlassoup)
  - Added prune juice (Daibousher)